ok so it was consts for whatever reason. This may require a new backend module, but we will see..
