+++ backend/src/utilities/file_utils.ts
@@ -1,37 +1,16 @@
 import {Request, Response, NextFunction, Router, RequestHandler} from "express";
-import multer from "multer";
-import path from "path";
-
-const storage = multer.diskStorage({
-            destination: (req,file,cb) => {
-                let dirname = `./uploads/sites/${req.params.site_id}`
-                cb(null,dirname)
-            },
-            filename: (req,file,cb) => {
-                let filename = `${req.params.point_id}${path.extname(file.originalname)}`
-                cb(null,filename);
-            }
-    })
-
-const do_upload = multer({
-    storage: storage,
-}).single("file");
 
 function upload_one(router: Router,
                     authenticate: (req: Request, res: Response, next: NextFunction) => Promise<any>,
+                    do_upload: RequestHandler<any>,
                     endpoint: string,
                     entity_name?: string){
     router.post(endpoint,
         authenticate,
+        do_upload,
         (req:Request,res:Response)=>{
-            do_upload(req,res,err => {
-                if(err){
-                    console.error(err)
-                    res.sendStatus(500);
-                } else {
-                    res.sendStatus(200);
-                }
-            })
+            //do_upload(req,res,err => {
+            //})
             console.log("THis should just work no?")
         })
 }
